diagnoser_starter/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
python-dotenv>=1.0
Werkzeug>=3.0
itsdangerous>=2.2
jinja2>=3.1
